--- packages/mp-loader/src/adapters/__tests__/__transpileModules__/key.js
@@ -1,19 +1,0 @@
-const { transformNode } = require('../../alipay/transpileModules/key');
-
-describe('Transpile module: key', () => {
-  it('should transform static key', () => {
-    const el = {
-      key: 'static-key',
-    };
-    transformNode(el);
-    expect(el.key).toEqual('static-key');
-  });
-
-  it('should transform key binded key', () => {
-    const el = {
-      key: 'tabs-index-{{idx}}',
-    };
-    transformNode(el);
-    expect(el.key).toEqual('("tabs-index-") + (data[\'idx\'])');
-  });
-});
