--- packages/mp-loader/src/adapters/alipay/transpileModules/bind.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/bind');
-
-const IS_DETECTIVE = /^a\:/;
-
-module.exports = {
-  transformNode: createTransformNode(IS_DETECTIVE)
-};
