--- packages/mp-loader/src/adapters/alipay/transpileModules/condition.js
@@ -1,11 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/condition');
-
-const ATTRIBUTES = {
-  IF: 'a:if',
-  ELSE_IF: 'a:elif',
-  ELSE: 'a:else'
-};
-
-module.exports = {
-  transformNode: createTransformNode(ATTRIBUTES)
-};
