--- packages/mp-loader/src/adapters/alipay/transpileModules/events.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/events');
-
-const LISTENER_ACTION = 'on';
-
-module.exports = {
-  transformNode: createTransformNode(LISTENER_ACTION)
-};
