--- packages/mp-loader/src/adapters/alipay/transpileModules/key.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/key');
-
-const ATTR_KEY = 'a:key';
-
-module.exports = {
-  transformNode: createTransformNode(ATTR_KEY)
-};
