--- packages/mp-loader/src/adapters/alipay/transpileModules/web-view.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/web-view');
-
-const MESSAGE_ATTRIBUTE = 'onMessage';
-
-module.exports = {
-  transformNode: createTransformNode(MESSAGE_ATTRIBUTE)
-};
