--- packages/mp-loader/src/adapters/common/transpileModules/attrs.js
@@ -1,5 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/attrs');
-
-module.exports = {
-  transformNode: createTransformNode()
-};
