--- packages/mp-loader/src/adapters/common/transpileModules/checked.js
@@ -1,5 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/checked');
-
-module.exports = {
-  transformNode: createTransformNode()
-};
