--- packages/mp-loader/src/adapters/common/transpileModules/template.js
@@ -1,6 +1,0 @@
-const { createTransformNode, genData } = require('../../../transpiler/transformCreator/template');
-
-module.exports = {
-  transformNode: createTransformNode(),
-  genData
-};
