--- packages/mp-loader/src/adapters/weapp/transpileModules/bind.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/bind');
-
-const IS_DETECTIVE = /^wx\:/;
-
-module.exports = {
-  transformNode: createTransformNode(IS_DETECTIVE)
-};
