--- packages/mp-loader/src/adapters/weapp/transpileModules/condition.js
@@ -1,11 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/condition');
-
-const ATTRIBUTES = {
-  IF: 'wx:if',
-  ELSE_IF: 'wx:elif',
-  ELSE: 'wx:else'
-};
-
-module.exports = {
-  transformNode: createTransformNode(ATTRIBUTES)
-};
