--- packages/mp-loader/src/adapters/weapp/transpileModules/events.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/events');
-
-const LISTENER_ACTION = 'bind';
-
-module.exports = {
-  transformNode: createTransformNode(LISTENER_ACTION)
-};
