--- packages/mp-loader/src/adapters/weapp/transpileModules/key.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/key');
-
-const ATTR_KEY = 'wx:key';
-
-module.exports = {
-  transformNode: createTransformNode(ATTR_KEY)
-};
