--- packages/mp-loader/src/adapters/weapp/transpileModules/list.js
@@ -1,11 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/list');
-
-const ATTRIBUTES = {
-  FOR: 'wx:for',
-  FOR_ITEM: 'wx:for-item',
-  FOR_IDX: 'wx:for-index'
-};
-
-module.exports = {
-  transformNode: createTransformNode(ATTRIBUTES)
-};
