--- packages/mp-loader/src/adapters/weapp/transpileModules/web-view.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../../transpiler/transformCreator/web-view');
-
-const MESSAGE_ATTRIBUTE = 'bindmessage';
-
-module.exports = {
-  transformNode: createTransformNode(MESSAGE_ATTRIBUTE)
-};
