--- packages/mp-loader/src/transpiler/transformCreator/checked.js
@@ -1,24 +1,0 @@
-const { getAndRemoveAttr, addAttr } = require('../helpers');
-
-const BOOLEAN_ATTRS = ['checked', 'disabled', 'plain'];
-
-function createTransformNode() {
-  return function(el) {
-    BOOLEAN_ATTRS.forEach((key) => {
-      if (Object.hasOwnProperty.call(el.attrsMap, key)) {
-        const exp = getAndRemoveAttr(el, key);
-
-        /**
-         * <tag checked />
-         * exp will be ""
-         * it means { checked: true }
-         */
-        addAttr(el, key, exp === '' ? 'true' : exp);
-      }
-    });
-  };
-}
-
-module.exports = {
-  createTransformNode
-};
