+++ packages/mp-loader/src/transpiler/transformModules/__tests__/__files__/template.axml
