+++ packages/mp-loader/src/transpiler/transformModules/__tests__/key.js
@@ -1,0 +1,19 @@
+const { transformNode } = require('../alipay/key');
+
+describe('Transpile module: key', () => {
+  it('should transform static key', () => {
+    const el = {
+      key: 'static-key',
+    };
+    transformNode(el);
+    expect(el.key).toEqual('static-key');
+  });
+
+  it('should transform key binded key', () => {
+    const el = {
+      key: 'tabs-index-{{idx}}',
+    };
+    transformNode(el);
+    expect(el.key).toEqual('("tabs-index-") + (data[\'idx\'])');
+  });
+});
