+++ packages/mp-loader/src/transpiler/transformModules/alipay/key.js
@@ -1,0 +1,7 @@
+const { createTransformNode } = require('../../transformCreator/key');
+
+const ATTR_KEY = 'a:key';
+
+module.exports = {
+  transformNode: createTransformNode(ATTR_KEY)
+};
