+++ packages/mp-loader/src/transpiler/transformModules/alipay/list.js
@@ -1,0 +1,11 @@
+const { createTransformNode } = require('../../transformCreator/list');
+
+const ATTRIBUTES = {
+  FOR: 'a:for',
+  FOR_ITEM: 'a:for-item',
+  FOR_IDX: 'a:for-index'
+};
+
+module.exports = {
+  transformNode: createTransformNode(ATTRIBUTES)
+};
