+++ packages/mp-loader/src/transpiler/transformModules/weapp/bind.js
@@ -1,0 +1,7 @@
+const { createTransformNode } = require('../../transformCreator/bind');
+
+const IS_DETECTIVE = /^wx\:/;
+
+module.exports = {
+  transformNode: createTransformNode(IS_DETECTIVE)
+};
