+++ packages/mp-loader/src/transpiler/transformModules/weapp/events.js
@@ -1,0 +1,7 @@
+const { createTransformNode } = require('../../transformCreator/events');
+
+const LISTENER_ACTION = 'bind';
+
+module.exports = {
+  transformNode: createTransformNode(LISTENER_ACTION)
+};
