+++ packages/mp-loader/src/transpiler/transformModules/weapp/web-view.js
@@ -1,0 +1,7 @@
+const { createTransformNode } = require('../../transformCreator/web-view');
+
+const MESSAGE_ATTRIBUTE = 'bindmessage';
+
+module.exports = {
+  transformNode: createTransformNode(MESSAGE_ATTRIBUTE)
+};
